refactor(rollup): address code-review findings on the estimator OOM fix
- Extract a grainHashExpr() helper in buildGrainProbeQuery instead of inlining
  the hash(concat_ws(...)) recipe twice.
- Move useDebouncedValue into packages/ui/src/hooks/ (the home for reusable hooks)
  instead of defining it inside the dimensions view.
- Correct the probe comment/JSDoc: hashing is a per-key constant-factor memory cut
  (~8x), not a structural bound — peak still grows with the number of enabled dims
  (N+1 distinct sets), so a pathologically wide grain could still be heavy.
- RollupImpactPanel: show the "Sync billing data" hint ahead of the stale/loading
  spinner so an un-synced install doesn't flash a spinner over the real no-data state.
- Soften the estimateStale comment: a single handoff frame can still paint the prior
  estimate after the debounce window closes (useQuery's loading state lags one tick).
